requirements.txt
requests
beautifulsoup4
flask
flask_restful
#qunicorn
opencv-python
azure-cognitiveservices-vision-face